// src/app/core/interfaces/buttoninterface.ts
export enum statusClassButton {
    DEFAULT = '',
    BUTTON_BLACK_00 = 'app-btn--black-00',
    BUTTON_PRIMARY = 'app-btn--primary',
    BUTTON_DARK_BLUE = 'app-btn--dark-blue',
    BUTTON_WHATSAPP = 'app-btn--whatsapp',
    BUTTON_WHITE_OUTLINE = 'app-btn--white-outline',
    BUTTON_INVERSE = 'app-btn--inverse',
    BUTTON_GRADIENT = 'app-btn--gradient',
    BUTTON_GHOST = 'app-btn--ghost',
    BUTTON_DANGER = 'app-btn--danger'
}